task 4agent/utils/pdf_reader.py
from pypdf import PdfReader
import PyPDF2


def extract_pdf_text(file):
    pdf_reader = PdfReader(file)
    text = ""
    for page in pdf_reader.pages:
        content = page.extract_text()
        if content:
            text += content + "\n"
    return text
