task 4agent/requirements.txt
streamlit
PyPDF2
google-generativeai
openagents
